Initial commit
Semi-functioning module is already here, but there's no documentation (except for
what's contained in the comments throughout the code). I haven't worked on this
for quite a while, so I'm going to try to get the documenation and some tests
set up before cleaning up the code and then working on fixing/adding more
functionality.