Add help and tests for [MagicDbProp()] attribute
First draft for the about_DatabaseReporter_DbColumnProp help file. The Pester
tests will be cleaned up a little more (you can tell where I started creating
a new module for each It block; the tests before that need to be made that way,
too), and the help will obviously be polished some more.

I'll need to figure out what to name the attribute at some point. The
DatabaseReporter engine allows renaming (and there's a minor Pester test that
makes sure that's working), but might need to look into coming up with a better
default name for the attribute...