Add SqlMetadata example module
The SqlMetadata module serves two purposes right now: it's an example of a
module you can create with the DatabaseReporter engine, and it also has a
very early version of a very useful helper command, New-DbReaderCommandSkeleton,
that, when finalized, will probably become incredibly valuable for building
other modules. Its purpose is to take a SQL query and build a simple version
of a DbReaderCommand instance.

The module has other commands for discovering schema information, but they're
not really tested very well yet, and of course the documentation and testing
aren't where they should be yet.