Make sure only DbReaderCommand functions are exported
This held true as long as a valid DbReaderCommand was specified. If there was
an error, or if no function was defined, then the default behavior of exporting
all functions was occuring.

This is b/c DbReaderCommand calls Export-ModuleMember, and if it didn't make it
through a successful call (either b/c there was a syntax error or b/c there was
no definition), then it wasn't called, and modules default to exporting all
functions if there's no PSD1 defining the exports and if Export-ModuleMember
wasn't called.

All it took to fix it was to put an empty call to Export-Module member at the
beginning. As DbReaderCommands are executed during module import, their
functions are exported one at a time.