Remove requirement to have [MagicDbProp()] parameters
Not having parameters that allow filtering reduces the usefulness of the DB
reader commands, but allowing an empty param() block makes it much easier
to make a new command for testing purposes. If a DBReaderCommand is created
with an empty param() block (or with a param() block that has no [MagicDbProp()]
attributes), then the SQL SELECT statement asks for all parameters.

This makes it so the engine doesn't have to check to make sure parameters have
been specified and write an error if they haven't.

I'm thinking of maybe testing out allowing arbitrary parameters in this case
so that you can feed the command whatever you want and have the WHERE clause
built up from that. That would allow a user to potentially pass arbitrary
strings to the SQL query, though (the parameter names would be passed into
the raw SQL statement). For now, I'm not even going to put it on the TODO list.

Allowing a 'SELECT *' meant that the InvokeReaderCommand needed to be updated,
too. It used to complain about duplicate column names, but now it ignores
duplicate names if the values are identical, and it creates (rather inefficiently),
unique column names if a duplicate name with different value is found.

Also updated the BasicFunctionality tests with a test for the 'SELECT *'.

While I was in there, I also added a test (that fails right now) that tests for
a known parameter conflict. This was mentioned in the TODO file, and needs to
be addressed ASAP.