Add test: [datetime] parameters should take $null
Noticed earlier that [datetime] parameters can't take $null. At first glance,
looks like DbReaderInfo transformation isn't added to the [datetime] parameters.
I can't remember if the DateTimeConverter helper is supposed to do this or not
(hopefully not...if so, that should be changed).

I'll dig into this tomorrow. Adding a call to AddDbReaderInfo in a different
argument transformation should fix this...