Update [datetime] test and fix NormalizeQuery
[datetime] test was importing the wrong DatabaseReporter.ps1 file, and the
test for not throwing wasn't sufficient to make sure a valid query was returned.
These issues are fixed.

Adding the parameter values as their own string when -ReturnSqlQuery is used was
causing issues with tests that checked the query. For this reason, NormalizeQuery
was set up to completely ignore the parameter string.