Fix FromClause parsing bug
FromClause currently implies the word 'FROM' being at the beginning. Defining
it should allow the word to exist, but it's not required. Before this commit,
a leading FROM would be deleted if the FromClause didn't have a leading space
before the start. If there was a leading space, though, the FROM wouldn't be
removed, and the final SQL query would have two FROMs, and it would be invalid.

This commit fixes that issue, while also making the parsing phase more flexible.

While looking at the issue, I found it odd that the parsing phase was happening
in a ScriptProperty that was being attached to the FromClause. I realized it was
designed this way so that a scriptblock could be passed, and if it had logic in
it dependent runtime conditions (like parameters on the function), it could
change what it outputs.

So I added a test to check for that, and it was failing because of the terrible
EvaluateAttributeArgumentValue function. It looks like this was an attempt to be
able to sanitize scriptblocks (and other AST tokens) during the parsing of the
DbReaderCommand phase. Uncommenting the call fixed the issue, and I think the
permanent answer is to remove it completely. Protecting against injection from
the end user is absolutely an important goal (and we're not there yet, even
though offering parameterized queries helps). Without that, modules that depend
on constrained endpoints to hide credentials (and the view of the DB the
authenticating credentials have access to) wouldn't be possible. If there's any
place a user can put dynamic content into the query that the command author
didn't anticipate, that's a problem. But, the command author should have the
flexibility to do things with late evaluated scriptblocks. Having a dynamic
FromClause is incredibly valuable, and if this function were hidden behind a
constrained endpoint, the end user shouldn't be able to alter it. So, FromClause
will happily work with scriptblocks for now, and EvaluateAttributeArgumentValue
will be looked at later to see if it should be completely removed, or if it
should be made better (and faster).

Also, EvaluateAttributeArgumentValue had extra Write-* and old commented out
code removed.