Fix module prefix bug
DbReaderCommand instances all share the same reference scriptblock. Before this
commit, that scriptblock would use $MyInvocation to determine its command name
and lookup that command's metadata from the module scoped $__CommandDeclarations
variable. That worked great unless the module had a prefix added at runtime.

Trying to use the .Module property wasn't much help since it wouldn't consistently
show the prefix that was in use. To fix the issue, the reference scripblock now
adds the hardcoded original command name in the begin {} block as the commands
definition is being created. (I lied earlier when I said they all shared the same
scriptblock...they all start from the same scriptblock, but the param() block was
already being assigned differently; now the first statement in the begin {} block
is, too).

This required a change in the way the begin, process, and end blocks were being
written out to the commands definition. The new way gets a chance to look at each
line, so we can do some filtering so that not all lines from the reference script-
block make it into the final command. For example, Write-Verbose and Write-Debug
statements won't make it unless $DebugMode is set to $true. We may look into having
some sort of tag at the end of a line to designate when a line shouldn't be
included in the final command (like right now when $DebugMode is $false, there are
extra foreach{} blocks that don't do anything since the Write-Verbose commands are
missing; it would be nice to have a way to not put that line in if debug mode is
off; then we also could be more granular about which Write-Verbose statements we
want to leave in)