Rename private variables in DbReaderCommand ref scriptblock
Renamed the following variables that were in the reference scriptblock:
  * Param
  * JoinSpacingString
  * SqlQuerySb
  * StringList
  * GroupByList
  * Property
  * CurrentSelect
  * ValidOrderByParameterNames
  * CombinedDbReaderInfo
  * OrderByStrings
  * SqlQuery
  * MaxWidth
  * ParamStrings
  * Entry
  * ConnectionParams

All now start with a '__'. The problem this solves is if an author creates a
parameter in the param() block with one of those names, it would impact the
behavior of the entire function. By putting underscores in the name, it should
make it so there can't be a chance of an accidental overwriting of a variable
this way.