Fix parameters that accept pipeline input
Last commit introduced a bug where pipeline input wasn't treated as a bound DB
column parameter. Issue occurred b/c variables were renamed, but there was one
variable name defined in the module scope: $__DbReaderInfoTableName is a module
scoped variable that holds the name of the $__PSBoundDbInfos. It was created in
the module scope just for the case that the last commit handled: renaming
variables. The __PSBoundDbInfos variable is modified in the argument
transformation scriptblock, which occurs inside the function scope when it's
executed.

I'm actually a little surprised everything worked when no pipeline input was
used. This part of the code was written a long time ago, so I'm hazy on all the
specifics, but it seems that more should have broken when the issue was
introduced. Need to refamiliarize myself with this and document how all of that
works.

Tests for ValueFromPipeline and ValueFromPipelineByPropertyName created, too.