Test GitHub documentation
PowerShell style help isn't going to work for documentation for this project
since the help files would never be loaded in the system. Better way to do this
would be to have the documenation online somewhere for easy viewing, but I'm
not about to try to keep up with that if it's separate from the code of the
project. Looks like GitHub has already addressed this with Github Pages, so I'm
trying to work through their instructions to see what this looks like. If it
works, I'll start converting the (lacking) existing documentation over.