Create tests for testing var and cmd polution
Some work is about to be done to make it so any module(s) created with
the DBR framework don't see unnecessary helper variables and commands
that the framework uses.

The test can't pass right now, but it works by taking a snapshot of
vars and cmds before dot sourcing the framework, then taking another
snapshot afterwards. From inside the module's scope, only excepted
variables and commands should be added (the excepted list is empty,
which is why the test can't pass right now)

Before making changes, here's a list of the vars and cmds that are
visible inside of a module that uses the framework:
  * Variable: __CommandDeclarations
  * Variable: __CommandsBefore
  * Variable: __DbReaderInfoTableName
  * Variable: __FakeAttributes
  * Variable: __OutputNullReplacementString
  * Variable: BoundDateTimeCompleter
  * Variable: BoundStandardCompleter
  * Variable: ReferenceCommandScriptBlock
  * Variable: StandardArgumentCompleter
  * Variable: TabExpansionAvailable
  * Command: AddDbReaderInfo
  * Command: AsDbConnectionType
  * Command: CombineDbReaderInfo
  * Command: ContainsMatch
  * Command: ConvertToSimpleXml
  * Command: DateTimeConverter
  * Command: DbReaderCommand
  * Command: EvaluateAttributeArgumentValue
  * Command: Get-DbReaderConnection
  * Command: GetInheritedClasses
  * Command: GetModuleSettingsDictionary
  * Command: GetValueFromAst
  * Command: InvokeReaderCommand
  * Command: MergeHelpInfo
  * Command: NewCompletionResult
  * Command: NewDbReaderInfo
  * Command: NewOrderByString
  * Command: New-TableFormatXml
  * Command: ParseParamBlock
  * Command: PrepReaderInfoForSql
  * Command: Set-DbReaderConnection