WIP: Fix scoping issues with DbReaderCommand being located in caller?
DbReaderCommand lives in the calling module's scope, not inside the DBR
dynamic module. (I'm cheating and re-doing this commit's text several
months after the fact, but I think this is to simplify scoping scenarios
with other side commands, and it might actually be necessary to get
Export-ModuleMember to export the defined commands)

Now that lots of important stuff lives inside the dynamic module, code
had to be added to smuggle that info from the module's internal scope.