Make parameters more generic
When testing, I was only using SQL Server, and was using parameters
that won't work for other DB technologies. SQLite was the first one that
I realized it didn't work with, and this update attempts to fix that by
making sure the parameter type name isn't hardcoded, and that parameters
are created by using a method on the DB command object.