Add [MagicDbProp()] validation at command parse time
Before this, invalid values provided to [MagicDbProp()] wouldn't cause
trouble until runtime. It's better to have issues with commands brought
up once when the command is being parsed from the definition.

This update only tests for the 'TransformArgument' attribute, but it
provides a mechanism so that all attributes can be tested. Instead of
defining the known attributes as an array, a hash table is used. The
keys are used like the old array, and any optional values are used to
transform the value provided in the command's defintion via a
ForEach-Object call.

Right now, the NewDbReaderInfo helper function makes sure things like
ComparisonOperator are valid, which helps prevent SQL injection attacks
when a user provides a hashtable to a parameter at run time.

In the future, need to merge these types of requirements with that
helper function to ensure that validation is controlled by a single
source, instead of having one set of validation rules run during parse
time, and one during command run time (validation still needs to happen
both places, but it should be the same set of rules)