Move FromClause evaluation and formatting
There was a failing test for dynamic FROM clauses that used
scriptblocks. I'm guessing the test started failing when all of the
framework functions/variables were moved into the dyanmic module scope,
since FROM clause evaluation was happening inside a script property
that was attached to the command author's original input.

This update changes that to what's probably a better system, even if
there weren't any problems.

Basically, FROM formatting happens inside a helper function, and the
FROM clause is evaluated each time the command runs, inside the
reference scriptblock. This means any variables defined in the param()
block that we may want to be able to change the FROM clause contents
are visible (before I *think* the scriptproperty execution was happening
in a different scope, so those type of variables weren't visible.)