Add -ReturnSqlQueryNew temp parameter
Getting ready to test something where I need to inspect the parameters
from a built query. The -ReturnSqlQuery functionality has needed to be
reworked ever since parameterized properties were added (showing their
values in a comment was a really quick way to still be able to use the
parameter to see what was going on interactively).

Changing how it works is going to require a change to all tests that
depend on it, though, so I'm temporarily adding a -ReturnSqlQueryNew
parameter so that tests that know how to access the parameters can use
it in the mean time.

When it's used, -ReturnSqlQuery is ignored, and a PSObject with two
properties (Query and Parameters) is returned.

When I do actually make this the behavior of -ReturnSqlQuery and fix up
any tests that use it, I'll probably also make the default formatting
show it in a friendly way (it would be trivial to use list formatting to
show the parameters the way the current -ReturnSqlQuery switch is
showing them).