Add test for SQLite wildcard escaping
Wildcard functionality up to this point has only taken SQL Server rules
into account. That means if a string has a literal '_' or '%' in it,
they are escaped according to SQL Server rules, which encloses them in
square brackets (the framework actually uses the [WildcardPattern]
helper method 'ToWql()' to do that)

While testing with SQLite, it soon became apparent that it doesn't do
escaping this way. There will be more of a discussion of this topic
when the functionality is fixed, but this commit adds a test that fails
now and should pass after the framework is updated.