Add documentation for [MagicDbInfo()] attribute
Mostly added this to try to start figuring out how to handle wildcard
replacement. Just writing out how I imagine the feature would work, the
WildcardReplacementScriptblock property mentioned in the documentation
would handle this. I'll make sure to update the documenation with any
change to the expected behavior when it's implemented.