Implement WildcardReplacementScriptblock attribute property
[MagicDbInfo(WildcardReplacementScriptblock)] can now be defined.

First big change for this commit is the $DBRModule-scoped hashtable
named $DbCommandInfoAttributeProperties, which will house all of the
optional property names so that changing the user-facing name is simple.

When DbReaderCommand is setting up a command, that hashtable's keys are
looped through, and the command will either use the value a user
specified, or (more commonly) will look up a default value using the
'GetDefaultAttributeProperty' helper function. The helper function will
be capable of changing the default value based on the RDBMS in use
(there's currently no way to change the RDBMS, though). The value that
it gets from the user or looks up in the default table is tucked away
in the command's $__MyCommandInfo variable.

Something that's currently missing is doing some sort of type
validation. Not sure of the best way to do that yet (there's so much
indirection spread all over this thing that it already makes my head
spin)

Finally, the 'WildcardStringToQuery' helper function now looks at the
scriptblock that was placed in the $__MyCommandInfo variable of the
DbReaderCommand.