Add SqlMode and WhereConditionBuilderScriptblock properties
These properties exist on the MagicDbInfo() attribute, and were added to
address SQLite literal _ and % escaping in LIKE conditions. An example
would be searching for a string that starts with 'test_', resulting in
a user-provided search string of 'test_*'. SQL wildcards would replace
the '*' with a '%' like this 'test_%'. Since '_' is also a SQL wildcard,
that exact string would match something like this: test123 (notice no
underscore).

The framework was handling this by using the ToWql() method of the
[WildcardPattern] type, which would automatically escape _ and %
characters. That method would turn 'test_*' into 'test[_]%', which SQL
Server understands. SQLite, however, doesn't understand that escaping.
It wants an escape character to be defined via the 'ESCAPE' clause after
the LIKE condition.

In SQLite, 'test_*' should be turned into 'test\_%', with a full WHERE
condition looking like the following:
  WHERE Name LIKE 'test\_%' ESCAPE '\'

Note that the escape character doesn't have to be '\'

This commit introduces the way that the framework will try to handle
making this work for multiple SQL flavors.

A previous commit introduced the default attribute value functionality,
and this one adds that capability for the scriptblock that builds WHERE
conditions.

It also introduces the SqlMode property, which is a special MagicDbInfo
property that is checked and populated (if it was hardcoded, by the
command author, that is used, otherwise some simple detection logic is
run to determine which mode to operate in). Right now, there's only two
modes: SQLite and everything else (even if the SqlMode is a garbage
string)

When SQLite WHERE conditions are being built, the operator is tested to
see if it's LIKE, and if it is, 'ESCAPE' is added to the end of the
condition.

Unfortunately, the escape character isn't configurable right now, and
the wildcard replacement logic happens separately from the WHERE
condition building. Long term, I'd like to address both of these issues,
but things should work this way short term, especially since the only
current changes to WHERE condition logic revolve around LIKE and
wildcardreplacement. So for now, 'WhereConditionBuilderScriptblock' and
'WildcardReplacementScriptblock' default changes can easily happen
together, but as more exceptions are created, and more SQL variants are
handled, the current structure for query building will probably turn
into a nightmare...